API - refacto
