+++ fittrackee/workouts/workouts.py
@@ -22,12 +22,13 @@ from fittrackee.equipments.exceptions import (
     InvalidEquipmentException,
     InvalidEquipmentsException,
 )
-from fittrackee.equipments.models import Equipment
+from fittrackee.equipments.models import Equipment, WorkoutEquipment
 from fittrackee.equipments.utils import (
     SPORT_EQUIPMENT_TYPES,
     handle_equipments,
 )
 from fittrackee.oauth2.server import require_auth
+from fittrackee.reports.models import ReportActionAppeal
 from fittrackee.responses import (
     DataInvalidPayloadErrorResponse,
     DataNotFoundErrorResponse,
@@ -44,9 +45,8 @@ from fittrackee.users.models import User, UserSportPreference
 from fittrackee.utils import decode_short_id
 from fittrackee.visibility_levels import can_view
 
-from ..reports.models import ReportActionAppeal
 from .decorators import check_workout
-from .models import Sport, Workout, WorkoutEquipment, WorkoutLike
+from .models import Sport, Workout, WorkoutLike
 from .utils.convert import convert_in_duration
 from .utils.gpx import (
     WorkoutGPXException,
